Not testing with node anymore, rename module to js.
